pyproject.toml
[project]
name = "shopping-with-ai"
version = "0.1.0"
description = "AI shopping assistant: RAG + price lookup + direct answer"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
 "openai>=1.40.0",
 "python-dotenv>=1.0.1",
 "faiss-cpu>=1.7.4",
 "numpy>=1.26.0",
 "langgraph>=0.2.0",
 "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
 "pytest>=8.0.0",
 "ruff>=0.6.0",
]

[tool.ruff]
line-length =100

[tool.pytest.ini_options]
testpaths = ["tests"]
